requirements.txt
asgiref==3.5.2
Django==3.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
python-slugify==7.0.0
pytz==2022.6
sqlparse==0.4.3
text-unidecode==1.3
